First PDB sequence gen
The script can be used to generate the sequence for the PDB file, which can be used for the template of a homollogy model. Specifically Modeller. The sequence is output in the FASTA format, ready for alignment.

It first downloads a dataset directly from PDB Http services. It is a FASTA formatted dataset which contains information about every entry in the Protein Data Bank: their sequences; and the residues which have not been observed experiemntally. These are not read by Modeller and need to be removed.

The function AliFinder() simply isolates the sequence, creates an index o fthe unobserved residues, and then removes them. Outputs the sequence as a FASTA file in the current working directory.

### Usage ###

AliFinder('ID:CHAIN')

in the above format. all caps.